Rename utilities file
