+++ lib/xiaomiTrvUtils.js
